Remove duplicate util function and move around sourcemap
